requirements.txt
streamlit==1.28.1
opencv-python==4.8.0.76
scikit-learn==1.3.2
scikit-image==0.22.0
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
pyyaml==6.0
requests==2.31.0
python-dotenv==1.0.0